requirements.txt
Authlib==1.2.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
decorator==5.1.1
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
requests==2.28.2
tqdm==4.65.0
urllib3==1.26.15
validators==0.20.0
weaviate-client==3.19.1
Werkzeug==2.3.4
